Move interpolation routines from fill.py to utils.py
