+++ roms_tools/setup/fill.py
@@ -419,128 +419,3 @@ def _iterative_fill_sor(nlat, nlon, var, fillmask, tol, rc, max_iter, fillvalue=
         iter_cnt += 1
 
     return var
-
-
-def interpolate_from_rho_to_u(field, method="additive"):
-
-    """
-    Interpolates the given field from rho points to u points.
-
-    This function performs an interpolation from the rho grid (cell centers) to the u grid
-    (cell edges in the xi direction). Depending on the chosen method, it either averages
-    (additive) or multiplies (multiplicative) the field values between adjacent rho points
-    along the xi dimension. It also handles the removal of unnecessary coordinate variables
-    and updates the dimensions accordingly.
-
-    Parameters
-    ----------
-    field : xr.DataArray
-        The input data array on the rho grid to be interpolated. It is assumed to have a dimension
-        named "xi_rho".
-
-    method : str, optional, default='additive'
-        The method to use for interpolation. Options are:
-        - 'additive': Average the field values between adjacent rho points.
-        - 'multiplicative': Multiply the field values between adjacent rho points. Appropriate for
-          binary masks.
-
-    Returns
-    -------
-    field_interpolated : xr.DataArray
-        The interpolated data array on the u grid with the dimension "xi_u".
-    """
-
-    if method == "additive":
-        field_interpolated = 0.5 * (field + field.shift(xi_rho=1)).isel(
-            xi_rho=slice(1, None)
-        )
-    elif method == "multiplicative":
-        field_interpolated = (field * field.shift(xi_rho=1)).isel(xi_rho=slice(1, None))
-    else:
-        raise NotImplementedError(f"Unsupported method '{method}' specified.")
-
-    if "lat_rho" in field_interpolated.coords:
-        field_interpolated.drop_vars(["lat_rho"])
-    if "lon_rho" in field_interpolated.coords:
-        field_interpolated.drop_vars(["lon_rho"])
-
-    field_interpolated = field_interpolated.swap_dims({"xi_rho": "xi_u"})
-
-    return field_interpolated
-
-
-def interpolate_from_rho_to_v(field, method="additive"):
-
-    """
-    Interpolates the given field from rho points to v points.
-
-    This function performs an interpolation from the rho grid (cell centers) to the v grid
-    (cell edges in the eta direction). Depending on the chosen method, it either averages
-    (additive) or multiplies (multiplicative) the field values between adjacent rho points
-    along the eta dimension. It also handles the removal of unnecessary coordinate variables
-    and updates the dimensions accordingly.
-
-    Parameters
-    ----------
-    field : xr.DataArray
-        The input data array on the rho grid to be interpolated. It is assumed to have a dimension
-        named "eta_rho".
-
-    method : str, optional, default='additive'
-        The method to use for interpolation. Options are:
-        - 'additive': Average the field values between adjacent rho points.
-        - 'multiplicative': Multiply the field values between adjacent rho points. Appropriate for
-          binary masks.
-
-    Returns
-    -------
-    field_interpolated : xr.DataArray
-        The interpolated data array on the v grid with the dimension "eta_v".
-    """
-
-    if method == "additive":
-        field_interpolated = 0.5 * (field + field.shift(eta_rho=1)).isel(
-            eta_rho=slice(1, None)
-        )
-    elif method == "multiplicative":
-        field_interpolated = (field * field.shift(eta_rho=1)).isel(
-            eta_rho=slice(1, None)
-        )
-    else:
-        raise NotImplementedError(f"Unsupported method '{method}' specified.")
-
-    if "lat_rho" in field_interpolated.coords:
-        field_interpolated.drop_vars(["lat_rho"])
-    if "lon_rho" in field_interpolated.coords:
-        field_interpolated.drop_vars(["lon_rho"])
-
-    field_interpolated = field_interpolated.swap_dims({"eta_rho": "eta_v"})
-
-    return field_interpolated
-
-
-def extrapolate_deepest_to_bottom(field: xr.DataArray, dim: str) -> xr.DataArray:
-    """
-    Extrapolate the deepest non-NaN values to the bottom along a specified dimension.
-
-    Parameters
-    ----------
-    field : xr.DataArray
-        The input data array containing NaN values that need to be filled. This array
-        should have at least one dimension named by `dim`.
-    dim : str
-        The name of the dimension along which to perform the interpolation and extrapolation.
-        Typically, this would be a vertical dimension such as 'depth' or 's_rho'.
-
-    Returns
-    -------
-    field_interpolated : xr.DataArray
-        A new data array with NaN values along the specified dimension filled by nearest
-        neighbor interpolation and extrapolation to the bottom. The original data array is not modified.
-
-    """
-    field_interpolated = field.interpolate_na(
-        dim=dim, method="nearest", fill_value="extrapolate"
-    )
-
-    return field_interpolated
